--- src/app/components/layout/Footer.tsx
@@ -1,20 +1,0 @@
-import { Box } from "@mui/material"
-
-export default function Footer() {
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        height: "100%",
-        flexDirection: "column",
-        justifyContent: "center",
-        alignItems: "center",
-        backgroundColor: "black",
-      }}
-    >
-      <span>dhdhdhdh</span>
-      <span>2222</span>
-      <span>dhdhdhdh</span>
-    </Box>
-  )
-}
